별점 조회 시 querydsl을 사용하도록 리팩토링

+++ src/main/java/ssu/eatssu/domain/review/dto/MealReviewsResponse.java
@@ -17,7 +17,7 @@ public class MealReviewsResponse implements ReviewInformationResponse {
     private List<String> menuNames;
 
     @Schema(description = "리뷰 개수", example = "15")
-    private Integer totalReviewCount;
+    private Long totalReviewCount;
 
     @Schema(description = "평점-메인", example = "4.4")
     private Double mainRating;
@@ -31,16 +31,16 @@ public class MealReviewsResponse implements ReviewInformationResponse {
     @Schema(description = "평점 별 갯수")
     private ReviewRatingCount reviewRatingCount;
 
-    public static MealReviewsResponse of(Integer totalReviewCount, List<String> menuNames,
-        RatingAverages ratingAverages, ReviewRatingCount reviewRatingCount) {
+    public static MealReviewsResponse of(Long totalReviewCount, List<String> menuNames,
+                                         RatingAverages ratingAverages, ReviewRatingCount reviewRatingCount) {
 
         return MealReviewsResponse.builder()
-            .menuNames(menuNames)
-            .mainRating(ratingAverages.mainRating())
-            .amountRating(ratingAverages.amountRating())
-            .tasteRating(ratingAverages.tasteRating())
-            .totalReviewCount(totalReviewCount)
-            .reviewRatingCount(reviewRatingCount)
-            .build();
+                .menuNames(menuNames)
+                .mainRating(ratingAverages.mainRating())
+                .amountRating(ratingAverages.amountRating())
+                .tasteRating(ratingAverages.tasteRating())
+                .totalReviewCount(totalReviewCount)
+                .reviewRatingCount(reviewRatingCount)
+                .build();
     }
 }
